Explicitly target a deployment or a statefulset
The most common source of invalid PDBs is the users creating a
deployment or a statefulset and a cronjob with the same application and
then simply using an application=foo selector. We can add a quick
workaround that will explicitly target pods belonging to a deployment or
a statefulset to avoid this situation.

Signed-off-by: Alexey Ermakov <alexey.ermakov@zalando.de>